Allow supplying of a SSH private key file
Functions in this role will autodiscover a key file to use if a password
has not been provided, but do not allow supplying a private key file to
use for authentication, despite py-junos-eznc supporting such an option.
Modify all the functions to allow providing a SSH private key file to be
used for authentication. Above said key must not be encrypted however as
py-junos-eznc does not seem to support asking for the passphrase to
unlock the private key nor does it seem to support ssh-agent